Move UI scripts to UI folder, refactor function names for consistency
