+++ Assets/Scripts/UI/SetButtonText.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5cc2b214052b53345a385d0db6003fca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
