+++ Assets/Scripts/UI/SimulationRateDropdown.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1432a4bdb7cb07b4eb6e147374f64ba6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
