+++ Assets/Scripts/UI/TextFromSlider.cs.meta
